+++ pkg/features/options.go
@@ -4,16 +4,9 @@ import "github.com/stackrox/rox/pkg/buildinfo"
 
 type option func(*feature)
 
-// Deprecated: once the feature needs to be enabled by default, the flag must be removed.
 func withEnabledByDefault() option {
 	return func(f *feature) {
 		f.defaultValue = true
-	}
-}
-
-func withTechPreviewStage() option {
-	return func(f *feature) {
-		f.techPreview = true
 	}
 }
 
@@ -24,7 +17,6 @@ func withUnchangeable(unchangeable bool) option {
 }
 
 var (
-	enabled            = withEnabledByDefault() // Deprecated: just remove the flag.
-	techPreview        = withTechPreviewStage()
+	enabled            = withEnabledByDefault()
 	unchangeableInProd = withUnchangeable(buildinfo.ReleaseBuild)
 )