+++ CHANGELOG.md
@@ -13,13 +13,16 @@ Please avoid adding duplicate information across this changelog and JIRA/doc inp
 
 - ROX-25066: Add new external backup integration for non-AWS S3 compatible providers.
 - ROX-25376: Add the release stage property to the `/v1/featureflags` response.
-- ROX-25451: Secured Cluster Auto-Upgrader is now enabled for all kind of clusters.
+- ROX-25451: Secured Cluster Auto-Upgrader is now enabled for all kinds of clusters.
+
 
 ### Removed Features
 
 ### Deprecated Features
 
 ### Technical Changes
+- ROX-24897: Sensor will now perform TLS checks lazily during delegated scanning instead of when secrets are first discovered, this should reduce Sensor startup time.
+  - To revert back to synchronous TLS checks set `ROX_SENSOR_LAZY_TLS_CHECKS` to `false` on Sensor.
 
 ## [4.5.0]